DMPlex: Generate halos from distributed DMPlex
Derives send and reordered receive lists for L1 and L2 node halos
from the DMPlex's point-SF and the previously generated halo region
label. We then use the same routines to generate the according
element halo (L1==L2) and let Fluidity derive ownership and
GNN-to-UNN maps.

Note: This requires level-2 overlap in DMPlexDistribute,
which is only avalibale in very recent versions of petsc.